Create layout with navbar based on the task below
Task1:

Position elements

Create a navigation bar template. Make a div element that stays at the top of the window wherever you scroll. Make sure that nothing slides behind the element on top of the page. Create an h1 with the content Layouts and make it centered in the HTML element.

1. There is a div tag in the HTML file containing an h1 tag

2. The div tag stays on top even when scrolling down the page

3. There is no element in the top of the page that is covered by the element.

4. The h1 display the text Layouts.

5. The Layouts text is aligned to the middle.